DRTVWR-493 Do not recreate proxy only to destroy it

+++ indra/llmessage/llproxy.cpp
@@ -404,8 +404,11 @@ LLSocks5AuthType LLProxy::getSelectedAuthMethod() const
 //static
 void LLProxy::cleanupClass()
 {
-	getInstance()->stopSOCKSProxy();
-	deleteSingleton();
+    if (instanceExists())
+    {
+        getInstance()->stopSOCKSProxy();
+        deleteSingleton();
+    }
 }
 
 void LLProxy::applyProxySettings(LLCurlEasyRequest* handle)
